requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch>=2.6.0
torchvision>=0.21.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
google-generativeai>=0.3.0
Werkzeug==3.0.1

